servers/python/requirements.txt
mcp-cli
chromadb
sentence-transformers
uv